feat(list): create a categories for mooods
