+++ src/components/Mood/moods.list.js
@@ -1,0 +1,174 @@
+const moods = [
+    "Abatid@",
+    "Abiert@",
+    "Aburrid@",
+    "Acongojad@",
+    "Afectad@",
+    "Afectuos@",
+    "Afligad@",
+    "Agitad@",
+    "Agobiad@",
+    "Agotad@",
+    "Agradecid@",
+    "Aislad@",
+    "Alarmad@",
+    "Alegre",
+    "Alentad@",
+    "Aliviad@",
+    "Amoros@",
+    "Angustiad@",
+    "Animad@",
+    "Añorad@",
+    "Ansios@",
+    "Apasionad@",
+    "Apátic@",
+    "Arrepetid@",
+    "Asco",
+    "Asustad@",
+    "Atemorizad@",
+    "Atent@",
+    "Aterrorizad@",
+    "Atraíd@",
+    "Avergonzad@",
+    "Bloquead@",
+    "Calmad@",
+    "Cansad@",
+    "Capaz",
+    "Cariños@",
+    "Celos@",
+    "Centrad@",
+    "Cercan@",
+    "Cómod@",
+    "Compasiv@",
+    "Complacid@",
+    "Comprensiv@",
+    "Comprometid@",
+    "Confiad@",
+    "Confus@",
+    "Conmocionad@",
+    "Conmovid@",
+    "Consternad@",
+    "Content@",
+    "Curios@",
+    "Débil",
+    "Decepcionad@",
+    "Decidid@",
+    "Deprimid@",
+    "Desahogad@",
+    "Desanimad@",
+    "Desbordad@",
+    "Descansad@",
+    "Desconcertad@",
+    "Desconectad@",
+    "Desconfiad@",
+    "Desconsolad@",
+    "Desdeños@",
+    "Deseos@",
+    "Desgraciad@",
+    "Deshech@",
+    "Desidios@",
+    "Desilusionad@",
+    "Desorientad@",
+    "Desprecio",
+    "Despreocupad@",
+    "Disgustad@",
+    "Dispuest@",
+    "Distante",
+    "Dividid@",
+    "Dolorid@",
+    "Dudos@",
+    "Efusiv@",
+    "Emocionad@",
+    "Enfadad@",
+    "Entusiasmad@",
+    "Envidios@",
+    "Escandalizad@",
+    "Escéptic@",
+    "Espectante",
+    "Esperanzad@",
+    "Estimulad@",
+    "Estresad@",
+    "Eufóric@",
+    "Exitad@",
+    "Exultante",
+    "Fascinad@",
+    "Fatigad@",
+    "Feliz",
+    "Frágil",
+    "Frí@",
+    "Frustrad@",
+    "Fuerte",
+    "Furios@",
+    "Generos@",
+    "Gozos@",
+    "Horrorizad@",
+    "Illusionad@",
+    "Illusionad@",
+    "Impaciente",
+    "Impotente",
+    "Impresionad@",
+    "Impresionad@",
+    "Incómod@",
+    "Indescis@",
+    "Indiferente",
+    "Indignad@",
+    "Infeliz",
+    "Inquiet@",
+    "Insegur@",
+    "Inspirad@",
+    "Inspirad@",
+    "Interesad@",
+    "Intranquil@",
+    "Intrigad@",
+    "Involucrad@",
+    "Irritable",
+    "Irritad@",
+    "Lúcid@",
+    "Malhumorad@",
+    "Maravillad@",
+    "Melancólic@",
+    "Molest@",
+    "Molest@",
+    "Motivad@",
+    "Nervios@",
+    "Nostálgic@",
+    "Odio",
+    "Optimist@",
+    "Orgullos@",
+    "Parad@",
+    "Pensativ@",
+    "Perdid@",
+    "Perplej@",
+    "Plácid@",
+    "Preocupad@",
+    "Preocupad@",
+    "Realizad@",
+    "Recepetiv@",
+    "Refrescad@",
+    "Relajad@",
+    "Renovad@",
+    "Reposad@",
+    "Reprochad@",
+    "Repugnad@",
+    "Resenti@",
+    "Restaurad@",
+    "Satisfech@",
+    "Segur@",
+    "Sensible",
+    "Simpátic@",
+    "Sobrecarg@",
+    "Sobrepasad@",
+    "Sobresaltad@",
+    "Sol@",
+    "Sorprendid@",
+    "Sorprendid@",
+    "Sosegad@",
+    "Tens@",
+    "Tiern@",
+    "Tímid@",
+    "Triste",
+    "Vacilante",
+    "Vulnerable",
+]
+
+export default moods;
